Rename plot lines/bands legend
