+++ components/CapacitiesAndInterventionsLegend.vue
